API Reference

Transfer Matrix Functions

Types

TransferMatrix.Layer — Type

Layer

A Layer stores information about a single layer, including its material name, thickness, a list of electric field wavelengths, and the real and imaginary parts of the refractive index associated with these wavelengths.

Initializing a Layer with no arguments makes a 1 μm thick layer of Air.

TransferMatrix.Structure — Type

Structure

The Structure is a mutable type that stores a Vector of Layer types, along with a list of field wavelengths and incident angles to calculate on.

Data Read/Write Functions

TransferMatrix.load_from_yaml — Function

read_yaml(file::String)

Read a yaml file with Structure parameters to perform transfer matrix calculations.

TransferMatrix.read_refractive — Method

read_refractive(f::String, material::String, thickness; div=1.0, freq=false)

Reads a csv file from refractiveindex.info containing a wavlength column in units of micrometers, and two refractive index columns, one for the real part and the other for the imaginary part.

The header names are:

"Wavelength, μm", "n", "k"

For some reason, the wavelength column cannot be normalized (so as to be a valid Julia identifier) using normalizednames = true in CSV.File, so the header is skipped.